restructure gui and working on hierarchy
