+++ core/compiler/examples/hierarchy.ar
@@ -1,0 +1,14 @@
+
+cell top() {
+  let left = bot()!;
+  let right = bot()!;
+  eq(left.x, 0);
+  eq(left.y, 0);
+  eq(right.x, 100);
+  eq(right.y, 0);
+  let met2 = rect(Layer::Met2, x0=left.via.x0, y0=20, x1=right.via.x1, y1=80)!;
+}
+
+cell bot() {
+    let via = rect(Layer::Met1, x0=0, y0=0, x1=100, y1=100)!;
+}
