+++ core/gui/src/editor/canvas.rs
