+++ core/gui/src/project.rs
@@ -5,6 +5,7 @@ use std::path::PathBuf;
 
 use compiler::compile::{CompileInput, CompiledCell, compile};
 use compiler::parse::parse;
+use compiler::solver::Var;
 use gpui::*;
 use itertools::Itertools;
 
@@ -16,131 +17,38 @@ use crate::{
     toolbars::{SideBar, TitleBar, ToolBar},
 };
 
-pub struct LayerState {
-    pub name: String,
-    pub color: Rgba,
-    pub fill: ShapeFill,
-    pub border_color: Rgba,
-    pub visible: bool,
-    pub z: usize,
+type Params = Vec<(String, f64)>;
+
+/// Identifier for specific parametrizations of p-cell with name `name`.
+struct CellId {
+    name: String,
+    params: Params,
 }
 
+/// Persistent state associated with a specific parametrization of a p-cell in a project.
+pub struct CellState {
+    pub rects: Vec<Rect<Var>>,
+    pub solved_values: Vec<(Var, f64)>,
+    // TODO: Use null space vectors to allow dragging coordinates.
+    pub null_space: (),
+    pub variable_overrides: Vec<(Var, f64)>,
+}
+
+/// Persistent state of project (i.e. anything that is saved in GUI project file).
+///
+/// GUI project file is saved in root directory of the associated Argon project.
 pub struct ProjectState {
-    pub path: PathBuf,
+    pub root: PathBuf,
     pub code: String,
-    pub cell: String,
-    pub params: HashMap<String, f64>,
-    pub solved_cell: CompiledCell,
-    pub rects: Vec<Rect>,
-    pub selected_rect: Option<usize>,
-    pub layers: Vec<Entity<LayerState>>,
-    pub subscriptions: Vec<Subscription>,
-    pub lsp_client: Option<GuiToLsp<TcpStream>>,
-}
-
-pub struct Project {
-    pub state: Entity<ProjectState>,
-    pub sidebar: Entity<SideBar>,
-    pub canvas: Entity<LayoutCanvas>,
-}
-
-fn get_rects(cx: &mut App, solved_cell: &CompiledCell, layers: &[Entity<LayerState>]) -> Vec<Rect> {
-    solved_cell
-        .values
-        .iter()
-        .filter_map(|v| v.get_rect().cloned())
-        .flat_map(|rect| {
-            let mut rects = Vec::new();
-            let layer = layers
-                .iter()
-                .map(|layer| (layer.clone(), layer.read(cx)))
-                .find(|(_, layer)| {
-                    if let Some(rect_layer) = &rect.layer {
-                        &layer.name == rect_layer
-                    } else {
-                        false
-                    }
-                });
-            if let Some((id, layer)) = layer {
-                rects.push(Rect {
-                    x0: rect.x0 as f32,
-                    y0: rect.y0 as f32,
-                    x1: rect.x1 as f32,
-                    y1: rect.y1 as f32,
-                    color: layer.color,
-                    fill: layer.fill,
-                    border_color: layer.border_color,
-                    layer: id.clone(),
-                    span: rect.source.clone().map(|info| info.span),
-                });
-            }
-            rects
-        })
-        .collect()
+    /// Specific parametrizations of p-cells that have been compiled.
+    pub cells: HashMap<CellId, CellData>,
+    /// Cells that are open in the GUI.
+    pub open_cells: Vec<CellId>,
+    pub lsp_client: GuiToLsp<TcpStream>,
 }
 
 impl Project {
-    pub fn new(cx: &mut Context<Self>, lsp_client: Option<GuiToLsp<TcpStream>>) -> Self {
-        let code = std::fs::read_to_string(&path).expect("failed to read file");
-        let ast = parse(&code).expect("failed to parse Argon");
-        let params_ref = params.iter().map(|(k, v)| (k.as_str(), *v)).collect();
-        let solved_cell = compile(CompileInput {
-            cell: &cell,
-            ast: &ast,
-            params: params_ref,
-        });
-        let layers: HashSet<_> = solved_cell
-            .values
-            .iter()
-            .filter_map(|value| value.get_rect()?.layer.clone())
-            .collect();
-        let layers: Vec<_> = layers
-            .into_iter()
-            .sorted()
-            .enumerate()
-            .map(|(z, name)| {
-                let mut s = DefaultHasher::new();
-                name.hash(&mut s);
-                let hash = s.finish() as usize;
-                let color = rgb([0xff0000, 0x0ff000, 0x00ff00, 0x000ff0, 0x0000ff][hash % 5]);
-                cx.new(|_cx| LayerState {
-                    name,
-                    color,
-                    fill: ShapeFill::Stippling,
-                    border_color: color,
-                    visible: true,
-                    z,
-                })
-            })
-            .collect();
-        let rects = get_rects(cx, &solved_cell, &layers);
-        let state = cx.new(|cx| {
-            let subscriptions = layers
-                .iter()
-                .map(|layer| {
-                    cx.observe(layer, |_, _, cx| {
-                        println!("project notified");
-                        cx.notify();
-                    })
-                })
-                .collect();
-            ProjectState {
-                path,
-                code,
-                cell,
-                params,
-                solved_cell: solved_cell.clone(),
-                rects,
-                selected_rect: None,
-                layers,
-                subscriptions,
-                lsp_client,
-            }
-        });
-
-        let sidebar = cx.new(|cx| SideBar::new(cx, &state));
-        let canvas = cx.new(|cx| LayoutCanvas::new(cx, &state));
-
+    pub fn new(cx: &mut Context<Self>, lsp_client: GuiToLsp<TcpStream>) -> Self {
         Self {
             state,
             sidebar,
